pyproject.toml
[project]
name = "rag-graph"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "faiss-cpu>=1.13.2",
    "fastapi>=0.135.2",
    "neo4j>=6.1.0",
    "openai>=2.30.0",
    "pytest>=9.0.2",
    "sentence-transformers>=5.3.0",
    "uvicorn>=0.42.0",
]
